+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Bump azurerm provider to 3.0.2.
+- Switch to Manual rolling mode for azure nodes.
 
 ## [8.2.0] - 2022-03-28